+++ src/gdb/GDBDebugSessionBase.ts
@@ -122,6 +122,8 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
     protected static frozenRequestArguments?: { request?: string };
 
     protected gdb!: IGDBBackend;
+    protected auxGdb?: IGDBBackend;
+
     protected isAttach = false;
     /**
      * True if GDB is using the "remote" or "extended-remote" target. In
@@ -247,6 +249,18 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
     }
 
     /**
+     * Validate the launch/attach request arguments and throw if they contain
+     * an invalid combination.
+     * @param args the request arguments to validate.
+     */
+    protected validateRequestArguments(
+        _args: LaunchRequestArguments | AttachRequestArguments
+    ) {
+        // No validation for 'gdb' debugger type. Derived classes may override.
+        // Added to base class for consistency.
+    }
+
+    /**
      * Apply the initial custom reset arguments.
      * @param args the arguments from the user to apply custom reset arguments to.
      */
@@ -331,7 +345,8 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
     }
 
     protected async setupCommonLoggerAndBackends(
-        args: LaunchRequestArguments | AttachRequestArguments
+        args: LaunchRequestArguments | AttachRequestArguments,
+        auxBackend?: boolean
     ) {
         logger.setup(
             args.verbose ? Logger.LogLevel.Verbose : Logger.LogLevel.Warn,
@@ -339,22 +354,32 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
         );
 
         const manager = await this.backendFactory.createGDBManager(this, args);
-        this.gdb = await this.backendFactory.createBackend(this, manager, args);
+        const gdb = await this.backendFactory.createBackend(
+            this,
+            manager,
+            args,
+            auxBackend ? 'AUX' : undefined
+        );
+        if (auxBackend) {
+            this.auxGdb = gdb;
+        } else {
+            this.gdb = gdb;
+        }
 
-        this.gdb.on('consoleStreamOutput', (output, category) => {
+        gdb.on('consoleStreamOutput', (output, category) => {
             const messageToPrint = this.switchOutputToError(output, category);
             this.sendEvent(
                 new OutputEvent(messageToPrint.output, messageToPrint.category)
             );
         });
 
-        this.gdb.on('execAsync', (resultClass, resultData) =>
+        gdb.on('execAsync', (resultClass, resultData) =>
             this.handleGDBAsync(resultClass, resultData)
         );
-        this.gdb.on('notifyAsync', (resultClass, resultData) =>
+        gdb.on('notifyAsync', (resultClass, resultData) =>
             this.handleGDBNotify(resultClass, resultData)
         );
-        this.gdb.on('resultAsync', (resultClass, resultData) =>
+        gdb.on('resultAsync', (resultClass, resultData) =>
             this.handleGDBResult(resultClass, resultData)
         );
     }
@@ -373,6 +398,7 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
         request: 'launch' | 'attach',
         args: LaunchRequestArguments | AttachRequestArguments
     ) {
+        this.validateRequestArguments(args);
         await this.setupCommonLoggerAndBackends(args);
 
         await this.spawn(args);
@@ -486,9 +512,10 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
     }
 
     protected async spawn(
-        args: LaunchRequestArguments | AttachRequestArguments
+        args: LaunchRequestArguments | AttachRequestArguments,
+        gdb: IGDBBackend = this.gdb
     ) {
-        return this.gdb?.spawn(args);
+        return gdb?.spawn(args);
     }
 
     /**
@@ -1694,17 +1721,26 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
     //     this.sendResponse(response);
     // }
 
+    /**
+     * Evaluates a GDB command.
+     * @param response Response object.
+     * @param expression Command to be executed, expected without leading '>'.
+     * @param frameRef Frame reference. If undefined, then it refers to the global scope.
+     */
     protected async evaluateRequestGdbCommand(
         response: DebugProtocol.EvaluateResponse,
-        args: DebugProtocol.EvaluateArguments,
+        expression: string,
         frameRef: FrameReference | undefined
     ): Promise<void> {
-        if (args.expression[1] === '-') {
-            await this.gdb.sendCommand(args.expression.slice(1));
+        const trimmedExpression = expression.trim();
+        if (trimmedExpression.startsWith('-')) {
+            // GDB/MI command
+            await this.gdb.sendCommand(trimmedExpression);
         } else {
+            // GDB CLI command
             await mi.sendInterpreterExecConsole(this.gdb, {
                 frameRef,
-                command: args.expression.slice(1),
+                command: trimmedExpression,
             });
         }
         response.body = {
@@ -1729,6 +1765,32 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
         return isInstructionBp;
     }
 
+    private async getFrameContext(
+        frameRef?: FrameReference
+    ): Promise<
+        [
+            IGDBBackend /* GDB Backend*/,
+            FrameReference | undefined /* Frame Reference */,
+            number /* Frame Depth */,
+            boolean /* isAux */,
+        ]
+    > {
+        if (this.auxGdb && this.isRunning) {
+            // Target is running and we have an auxiliary GDB connection,
+            // hence we don't have a stable stack to ask the depth for.
+            return [this.auxGdb, undefined, 0, true];
+        }
+        // Read stack depth from main GDB connection, in reality
+        // getFrameContext() is (or better should) be never called
+        // for a running target without auxiliary GDB connection.
+        const stackDepth = await mi.sendStackInfoDepth(this.gdb, {
+            maxDepth: 100,
+            threadId: frameRef?.threadId,
+        });
+        const depth = parseInt(stackDepth.depth, 10);
+        return [this.gdb, frameRef, depth, false];
+    }
+
     protected async evaluateRequest(
         response: DebugProtocol.EvaluateResponse,
         args: DebugProtocol.EvaluateArguments
@@ -1746,25 +1808,33 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
             variablesReference: 0,
         }; // default response
         try {
-            const allowCliCommand =
-                alwaysAllowCliCommand && args.expression.startsWith('>');
+            const expression = args.expression.trim();
+            const isCliCommand =
+                expression.startsWith('>') && args.context === 'repl';
+            const isAux = this.auxGdb && this.isRunning;
+            // Allow evaluation of expression without frame info when CLI commands are always allowed or
+            // when called with normal expression in auxiliary GDB mode.
+            const allowUndefinedFrame =
+                (isCliCommand && alwaysAllowCliCommand) ||
+                (!isCliCommand && isAux);
 
-            if (!allowCliCommand && args.frameId === undefined) {
+            if (!allowUndefinedFrame && args.frameId == undefined) {
                 throw new Error(
                     'Evaluation of expression without frameId is not supported.'
                 );
             }
 
-            const frameRef = args.frameId
+            const initialFrameRef = args.frameId
                 ? this.frameHandles.get(args.frameId)
                 : undefined;
 
-            if (!allowCliCommand && !frameRef) {
+            if (!allowUndefinedFrame && !initialFrameRef) {
                 this.sendResponse(response);
                 return;
             }
 
-            if (args.expression.startsWith('>') && args.context === 'repl') {
+            if (isCliCommand) {
+                const expressionNoPrefix = expression.slice(1).trim();
                 const regexDisable = new RegExp(
                     '^\\s*(?:dis|disa|disable)\\s*(?:(?:breakpoint|count|delete|once)\\d*)?\\s*\\d*\\s*$'
                 );
@@ -1775,8 +1845,8 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
                     '^\\s*(?:d|del|delete)\\s+(?:breakpoints\\s+)?(\\d+)?\\s*$'
                 );
                 if (
-                    args.expression.slice(1).search(regexDisable) != -1 ||
-                    args.expression.slice(1).search(regexEnable) != -1
+                    expressionNoPrefix.search(regexDisable) != -1 ||
+                    expressionNoPrefix.search(regexEnable) != -1
                 ) {
                     this.sendEvent(
                         new OutputEvent(
@@ -1785,9 +1855,7 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
                         )
                     );
                 }
-                const deleteRegexMatch = args.expression
-                    .slice(1)
-                    .match(regexDelete);
+                const deleteRegexMatch = expressionNoPrefix.match(regexDelete);
                 if (deleteRegexMatch) {
                     if (
                         await this.isInstructionBreakpoint(deleteRegexMatch[1])
@@ -1802,35 +1870,31 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
                 }
                 return await this.evaluateRequestGdbCommand(
                     response,
-                    args,
-                    frameRef
+                    expressionNoPrefix,
+                    initialFrameRef
                 );
             }
 
-            const stackDepth = await mi.sendStackInfoDepth(this.gdb, {
-                maxDepth: 100,
-            });
-            const depth = parseInt(stackDepth.depth, 10);
-            let varobj = this.gdb.varManager.getVar(
-                frameRef,
-                depth,
-                args.expression
-            );
+            const [gdb, frameRef, depth] =
+                await this.getFrameContext(initialFrameRef);
+
+            let varobj = gdb.varManager.getVar(frameRef, depth, expression);
             if (!varobj) {
-                const varCreateResponse = await mi.sendVarCreate(this.gdb, {
-                    expression: args.expression,
+                const varCreateResponse = await mi.sendVarCreate(gdb, {
+                    expression,
                     frameRef,
+                    frame: frameRef ? 'current' : 'floating',
                 });
-                varobj = this.gdb.varManager.addVar(
+                varobj = gdb.varManager.addVar(
                     frameRef,
                     depth,
-                    args.expression,
+                    expression,
                     false,
                     false,
                     varCreateResponse
                 );
             } else {
-                const vup = await mi.sendVarUpdate(this.gdb, {
+                const vup = await mi.sendVarUpdate(gdb, {
                     name: varobj.varname,
                 });
                 const update = vup.changelist[0];
@@ -1845,20 +1909,17 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
                             depth,
                             varobj.varname
                         );
-                        await mi.sendVarDelete(this.gdb, {
+                        await mi.sendVarDelete(gdb, {
                             varname: varobj.varname,
                         });
-                        const varCreateResponse = await mi.sendVarCreate(
-                            this.gdb,
-                            {
-                                expression: args.expression,
-                                frameRef,
-                            }
-                        );
-                        varobj = this.gdb.varManager.addVar(
+                        const varCreateResponse = await mi.sendVarCreate(gdb, {
+                            expression,
+                            frameRef,
+                        });
+                        varobj = gdb.varManager.addVar(
                             frameRef,
                             depth,
-                            args.expression,
+                            expression,
                             false,
                             false,
                             varCreateResponse
@@ -1866,22 +1927,24 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
                     }
                 }
             }
-            if (varobj && args.frameId != undefined) {
+            if (varobj) {
+                const frameHandle = args.frameId ?? -1;
                 const result =
                     args.context === 'variables' && Number(varobj.numchild)
-                        ? await this.getChildElements(varobj, args.frameId)
+                        ? await this.getChildElements(varobj, frameHandle)
                         : varobj.value;
+                const variablesReference =
+                    parseInt(varobj.numchild, 10) > 0
+                        ? this.variableHandles.create({
+                              type: 'object',
+                              frameHandle,
+                              varobjName: varobj.varname,
+                          })
+                        : 0;
                 response.body = {
                     result,
                     type: varobj.type,
-                    variablesReference:
-                        parseInt(varobj.numchild, 10) > 0
-                            ? this.variableHandles.create({
-                                  type: 'object',
-                                  frameHandle: args.frameId,
-                                  varobjName: varobj.varname,
-                              })
-                            : 0,
+                    variablesReference,
                 };
             }
 
@@ -2050,10 +2113,11 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
         response: DebugProtocol.ReadMemoryResponse,
         args: DebugProtocol.ReadMemoryArguments
     ): Promise<void> {
+        const gdb = this.auxGdb && this.isRunning ? this.auxGdb : this.gdb;
         try {
             if (args.count) {
                 const result = await mi.sendDataReadMemoryBytes(
-                    this.gdb,
+                    gdb,
                     args.memoryReference,
                     args.count,
                     args.offset
@@ -2082,6 +2146,8 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
         response: DebugProtocol.WriteMemoryResponse,
         args: DebugProtocol.WriteMemoryArguments
     ) {
+        const gdb = this.auxGdb && this.isRunning ? this.auxGdb : this.gdb;
+
         try {
             const { memoryReference, data } = args;
             const typeofAddress = typeof memoryReference;
@@ -2097,11 +2163,7 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
                 );
             }
             const hexContent = base64ToHex(data);
-            await mi.sendDataWriteMemoryBytes(
-                this.gdb,
-                memoryReference,
-                hexContent
-            );
+            await mi.sendDataWriteMemoryBytes(gdb, memoryReference, hexContent);
             this.sendResponse(response);
         } catch (err) {
             this.sendErrorResponse(
@@ -2506,6 +2568,14 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
     protected async handleVariableRequestFrame(
         ref: FrameVariableReference
     ): Promise<DebugProtocol.Variable[]> {
+        if (this.auxGdb && this.isRunning) {
+            // Don't allow (local) variables view when using an auxiliary GDB,
+            // (local) variables often can't be read while running.
+            throw new Error(
+                'Cannot handle variable frame requests while target is running'
+            );
+        }
+
         // initialize variables array and dereference the frame handle
         const variables: DebugProtocol.Variable[] = [];
         const frameRef = this.frameHandles.get(ref.frameHandle);
@@ -2561,7 +2631,7 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
                         let value = varobj.value;
                         // if we have an array parent entry, we need to display the address.
                         if (arrayRegex.test(varobj.type)) {
-                            value = await this.getAddr(varobj);
+                            value = await this.getAddr(varobj, this.gdb);
                         }
                         variables.push({
                             name: varobj.expression,
@@ -2624,7 +2694,7 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
                 let value = varobj.value;
                 // if we have an array parent entry, we need to display the address.
                 if (arrayRegex.test(varobj.type)) {
-                    value = await this.getAddr(varobj);
+                    value = await this.getAddr(varobj, this.gdb);
                 }
                 variables.push({
                     name: varobj.expression,
@@ -2649,37 +2719,36 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
     protected async handleVariableRequestObject(
         ref: ObjectVariableReference
     ): Promise<DebugProtocol.Variable[]> {
-        // initialize variables array and dereference the frame handle
+        // initialize variables array
         const variables: DebugProtocol.Variable[] = [];
-        const frameRef = this.frameHandles.get(ref.frameHandle);
-        if (!frameRef) {
+        // dereference the frame handle
+        const initialFrameRef = this.frameHandles.get(ref.frameHandle);
+        const [gdb, frameRef, depth, isAux] =
+            await this.getFrameContext(initialFrameRef);
+
+        if (!isAux && !frameRef) {
             return Promise.resolve(variables);
         }
 
-        // fetch stack depth to obtain frameId/threadId/depth tuple
-        const stackDepth = await mi.sendStackInfoDepth(this.gdb, {
-            maxDepth: 100,
-        });
-        const depth = parseInt(stackDepth.depth, 10);
         // we need to keep track of children and the parent varname in GDB
         let children;
         let parentVarname = ref.varobjName;
 
         // if a varobj exists, use the varname stored there
-        const varobj = this.gdb.varManager.getVarByName(
+        const varobj = gdb.varManager.getVarByName(
             frameRef,
             depth,
             ref.varobjName
         );
         if (varobj) {
-            children = await mi.sendVarListChildren(this.gdb, {
+            children = await mi.sendVarListChildren(gdb, {
                 name: varobj.varname,
                 printValues: mi.MIVarPrintValues.all,
             });
             parentVarname = varobj.varname;
         } else {
             // otherwise use the parent name passed in the variable reference
-            children = await mi.sendVarListChildren(this.gdb, {
+            children = await mi.sendVarListChildren(gdb, {
                 name: ref.varobjName,
                 printValues: mi.MIVarPrintValues.all,
             });
@@ -2687,7 +2756,7 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
         // Grab the full path of parent.
         const topLevelPathExpression =
             varobj?.expression ??
-            (await this.getFullPathExpression(parentVarname));
+            (await this.getFullPathExpression(parentVarname, gdb));
 
         // iterate through the children
         for (const child of children.children) {
@@ -2695,7 +2764,7 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
             const isClass = this.isChildOfClass(child);
             if (isClass) {
                 const name = `${parentVarname}.${child.exp}`;
-                const objChildren = await mi.sendVarListChildren(this.gdb, {
+                const objChildren = await mi.sendVarListChildren(gdb, {
                     name,
                     printValues: mi.MIVarPrintValues.all,
                 });
@@ -2758,9 +2827,12 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
     }
 
     /** Query GDB using varXX name to get complete variable name */
-    protected async getFullPathExpression(inputVarName: string) {
+    protected async getFullPathExpression(
+        inputVarName: string,
+        gdb: IGDBBackend
+    ) {
         const exprResponse = await mi.sendVarInfoPathExpression(
-            this.gdb,
+            gdb,
             inputVarName
         );
         // result from GDB looks like (parentName).field so remove ().
@@ -2774,6 +2846,12 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
     protected async handleVariableRequestRegister(
         ref: RegisterVariableReference
     ): Promise<DebugProtocol.Variable[]> {
+        if (this.auxGdb && this.isRunning) {
+            // Don't allow register view when using an auxiliary GDB, core registers
+            // often can't be read while running.
+            throw new Error('Cannot read registers while target is running');
+        }
+
         // initialize variables array and dereference the frame handle
         const variables: DebugProtocol.Variable[] = [];
         const frameRef = this.frameHandles.get(ref.frameHandle);
@@ -2821,9 +2899,9 @@ export abstract class GDBDebugSessionBase extends LoggingDebugSession {
         return Promise.resolve(variables);
     }
 
-    protected async getAddr(varobj: VarObjType) {
+    protected async getAddr(varobj: VarObjType, gdb: IGDBBackend) {
         const addr = await mi.sendDataEvaluateExpression(
-            this.gdb,
+            gdb,
             `&(${varobj.expression})`
         );
         return addr.value ? addr.value : varobj.value;