+++ src/web/GDBTargetDebugSession.ts
@@ -9,8 +9,8 @@
  *********************************************************************/
 
 import { GDBDebugSession } from './GDBDebugSession';
-import { logger, OutputEvent, TerminatedEvent } from '@vscode/debugadapter';
-import { LogLevel } from '@vscode/debugadapter/lib/logger';
+import { OutputEvent, TerminatedEvent } from '@vscode/debugadapter';
+import { logger, LogLevel } from '@vscode/debugadapter/lib/logger';
 import * as mi from '../mi';
 import { DebugProtocol } from '@vscode/debugprotocol';
 import {
@@ -135,6 +135,29 @@ export class GDBTargetDebugSession extends GDBDebugSession {
         }
     }
 
+    /**
+     * Validate the launch/attach request arguments and throw if they contain
+     * an invalid combination.
+     * @param args the request arguments to validate.
+     */
+    protected validateRequestArguments(
+        args: TargetLaunchRequestArguments | TargetAttachRequestArguments
+    ) {
+        if (args.auxiliaryGdb) {
+            // Limitations for auxiliary GDB mode
+            if (args.gdbNonStop) {
+                throw new Error(
+                    'Cannot use auxiliaryGdb mode with gdbNonStop mode'
+                );
+            }
+            if (args.gdbAsync === false) {
+                throw new Error(
+                    'AuxiliaryGdb mode requires gdbAsync to be active'
+                );
+            }
+        }
+    }
+
     protected override async setupCommonLoggerAndBackends(
         args: TargetLaunchRequestArguments | TargetAttachRequestArguments
     ) {
@@ -149,6 +172,7 @@ export class GDBTargetDebugSession extends GDBDebugSession {
         request: 'launch' | 'attach',
         args: TargetLaunchRequestArguments | TargetAttachRequestArguments
     ) {
+        this.validateRequestArguments(args);
         await this.setupCommonLoggerAndBackends(args);
         this.initializeSessionArguments(args);